requirements.txt
openai
psycopg2-binary
python-dotenv
pgvector
numpy
flask
flask-cors
scikit-learn
